requirements.txt
PyGObject>=3.42.0
yt-dlp>=2024.3.10
# TensorFlow – pick the build that matches the Mac architecture
tensorflow-macos>=2.15 ; platform_machine=="arm64" and sys_platform=="darwin"
tensorflow-metal>=1.1   ; platform_machine=="arm64" and sys_platform=="darwin"
# For Intel Macs (or Apple-silicon terminals running under Rosetta)
tensorflow>=2.15,<3     ; platform_machine=="x86_64" and sys_platform=="darwin"
pillow>=9.5.0
numpy>=1.23
fonttools>=4.50
librosa>=0.10
opencv-python>=4.8
tqdm>=4.66 